refactor: extract delegation executor module
