+++ packages/standalone/src/agent/gateway-tool-executor.ts
@@ -82,7 +82,6 @@ import {
   compareVersionMetrics,
   getActivity,
   logActivity,
-  updateActivityScore,
 } from '../db/agent-store.js';
 import type { RoleConfig } from '../cli/config/types.js';
 import { DEFAULT_ROLES } from '../cli/config/types.js';
@@ -94,7 +93,12 @@ import {
   validateManagedAgentCreateInput,
   validateManagedAgentChanges,
 } from './managed-agent-validation.js';
-import type { ValidationSessionRow } from '../validation/types.js';
+import {
+  DelegationExecutor,
+  type AgentTestInput,
+  type DelegateInput,
+  type DelegationRoutingContext,
+} from './delegation-executor.js';
 import { EnvelopeEnforcer, EnvelopeViolation } from '../envelope/index.js';
 import type { Envelope } from '../envelope/index.js';
 
@@ -163,20 +167,6 @@ function sanitizeCommandForAudit(command: string): { commandHash: string; comman
     .slice(0, 200);
 
   return { commandHash, commandPreview };
-}
-
-function summarizeActivityOutput(output: unknown): string | undefined {
-  if (output === undefined || output === null) {
-    return undefined;
-  }
-  if (typeof output === 'string') {
-    return output.slice(0, 500);
-  }
-  try {
-    return JSON.stringify(output).slice(0, 500);
-  } catch {
-    return String(output).slice(0, 500);
-  }
 }
 
 /**
@@ -292,12 +282,14 @@ export class GatewayToolExecutor {
   private sessionsDb: SQLiteDatabase | null = null;
   setSessionsDb(db: SQLiteDatabase): void {
     this.sessionsDb = db;
+    this.refreshDelegationExecutor();
   }
   private rawStore: import('../connectors/framework/raw-store.js').RawStore | null = null;
   setRawStore(store: import('../connectors/framework/raw-store.js').RawStore): void {
     this.rawStore = store;
+    this.refreshDelegationExecutor();
   }
-  private testInFlight = new Map<string, Promise<GatewayToolResult>>();
+  private delegationExecutor: DelegationExecutor | null = null;
   private uiCommandQueue: UICommandQueue | null = null;
   setUICommandQueue(queue: UICommandQueue): void {
     this.uiCommandQueue = queue;
@@ -317,15 +309,18 @@ export class GatewayToolExecutor {
     svc: import('../validation/session-service.js').ValidationSessionService
   ): void {
     this.validationService = svc;
+    this.refreshDelegationExecutor();
   }
   setMemoryAgent(processManager: AgentProcessManager): void {
     this.memoryAgentProcessManager = processManager;
   }
   setAgentProcessManager(pm: AgentProcessManager): void {
     this.agentProcessManager = pm;
+    this.refreshDelegationExecutor();
   }
   setDelegationManager(dm: DelegationManager): void {
     this.delegationManagerRef = dm;
+    this.refreshDelegationExecutor();
   }
   /** Get AgentProcessManager (for cron/event triggers that need direct process access) */
   getAgentProcessManager(): AgentProcessManager | null {
@@ -367,7 +362,7 @@ export class GatewayToolExecutor {
     return this.getExecutionState().agentContext;
   }
 
-  private getActiveRouting(): { agentId: string; source: string; channelId: string } {
+  private getActiveRouting(): DelegationRoutingContext {
     const state = this.getExecutionState();
     return {
       agentId: state.agentId,
@@ -399,29 +394,6 @@ export class GatewayToolExecutor {
   }
   setDisallowedGatewayTools(tools: string[]): void {
     this.disallowedGatewayTools = new Set(tools);
-  }
-
-  private cleanupValidationSessionOnTelemetryFailure(
-    session: ValidationSessionRow | null,
-    error: unknown,
-    label: string
-  ): null {
-    if (!session || !this.validationService) {
-      return null;
-    }
-    const message = error instanceof Error ? error.message : String(error);
-    try {
-      this.validationService.finalizeSession(session.id, {
-        execution_status: 'failed',
-        error_message: `${label}: ${message}`,
-      });
-    } catch (cleanupErr) {
-      securityLogger.warn(
-        `[Delegation telemetry] Failed to clean up validation session ${session.id}`,
-        cleanupErr
-      );
-    }
-    return null;
   }
 
   private getPreferredViewerAgentTab(): string {
@@ -535,6 +507,30 @@ export class GatewayToolExecutor {
   /** Retry delay (ms) for delegate backoff. Initialized from config in constructor. */
   private _retryDelayMs: number = 1000;
 
+  private createDelegationExecutor(): DelegationExecutor {
+    return new DelegationExecutor({
+      agentProcessManager: this.agentProcessManager,
+      delegationManagerRef: this.delegationManagerRef,
+      rawStore: this.rawStore,
+      sessionsDb: this.sessionsDb,
+      validationService: this.validationService,
+      retryDelayMs: this._retryDelayMs,
+      resolveManagedAgentId: (id) => this.resolveManagedAgentId(id),
+      checkViewerOnly: () => this.checkViewerOnly(),
+    });
+  }
+
+  private refreshDelegationExecutor(): void {
+    this.delegationExecutor = this.createDelegationExecutor();
+  }
+
+  private getDelegationExecutor(): DelegationExecutor {
+    if (!this.delegationExecutor) {
+      this.refreshDelegationExecutor();
+    }
+    return this.delegationExecutor!;
+  }
+
   constructor(options: GatewayToolExecutorOptions = {}) {
     this.mamaDbPath = options.mamaDbPath;
     this.sessionStore = options.sessionStore;
@@ -556,6 +552,7 @@ export class GatewayToolExecutor {
     } catch {
       // Config not initialized yet — keep default 1000ms
     }
+    this.refreshDelegationExecutor();
   }
 
   /**
@@ -903,12 +900,9 @@ export class GatewayToolExecutor {
           );
         // Agent lifecycle tools
         case 'agent_test':
-          return await this.executeAgentTest(
-            input as {
-              agent_id: string;
-              sample_count?: number;
-              test_data?: Array<{ input: string; expected?: string }>;
-            }
+          return await this.getDelegationExecutor().runAgentTest(
+            input as AgentTestInput,
+            this.getActiveRouting()
           );
         // Agent management tools (Managed Agents pattern)
         case 'agent_get': {
@@ -1144,8 +1138,9 @@ export class GatewayToolExecutor {
         }
         // Multi-Agent delegation
         case 'delegate':
-          return await this.executeDelegate(
-            input as { agentId: string; task: string; background?: boolean }
+          return await this.getDelegationExecutor().runDelegate(
+            input as DelegateInput,
+            this.getActiveRouting()
           );
       }
 
@@ -2672,672 +2667,6 @@ export class GatewayToolExecutor {
         error: `Failed to send to webchat: ${err instanceof Error ? err.message : String(err)}`,
       };
     }
-  }
-
-  // ============================================================================
-  // Multi-Agent Delegation
-  // ============================================================================
-
-  /**
-   * Execute delegate tool — dispatch a task to another agent
-   */
-  // ── Agent Test ─────────────────────────────────────────────────────────────
-
-  private async executeAgentTest(input: {
-    agent_id: string;
-    sample_count?: number;
-    test_data?: Array<{ input: string; expected?: string }>;
-  }): Promise<GatewayToolResult> {
-    const permError = this.checkViewerOnly();
-    if (permError) {
-      return { success: false, error: permError } as GatewayToolResult;
-    }
-
-    const { agent_id } = input;
-    const sample_count = Number.parseInt(String(input.sample_count ?? 2), 10);
-    const resolvedAgentId = this.resolveManagedAgentId(agent_id);
-    if (!Number.isFinite(sample_count) || sample_count < 1) {
-      securityLogger.warn('[Agent test] Invalid sample_count received', {
-        agent_id: resolvedAgentId,
-        sample_count: input.sample_count ?? null,
-      });
-      return {
-        success: false,
-        error: `Invalid sample_count for '${resolvedAgentId}': ${String(input.sample_count)}. Must be >= 1.`,
-      } as GatewayToolResult;
-    }
-
-    // Concurrency guard
-    if (this.testInFlight.has(resolvedAgentId)) {
-      return { success: false, error: 'test_already_running' } as GatewayToolResult;
-    }
-
-    const promise = this._runAgentTest(resolvedAgentId, sample_count, input.test_data);
-    this.testInFlight.set(resolvedAgentId, promise);
-    try {
-      return await promise;
-    } finally {
-      this.testInFlight.delete(resolvedAgentId);
-    }
-  }
-
-  private async _runAgentTest(
-    agentId: string,
-    sampleCount: number,
-    testData?: Array<{ input: string; expected?: string }>
-  ): Promise<GatewayToolResult> {
-    if (!this.agentProcessManager || !this.delegationManagerRef) {
-      return {
-        success: false,
-        error: 'agent_timeout: multi-agent not configured',
-      } as GatewayToolResult;
-    }
-
-    const startTime = Date.now();
-
-    // 1. Collect test data
-    let items: Array<{ input: string; expected?: string }>;
-    if (testData && testData.length > 0) {
-      const normalizedItems: Array<{ input: string; expected?: string }> = [];
-      for (let index = 0; index < testData.length; index++) {
-        const rawItem = testData[index] as unknown as Record<string, unknown>;
-        if (typeof rawItem.input !== 'string') {
-          return {
-            success: false,
-            error: `Invalid test_data[${index}].input: expected string`,
-          } as GatewayToolResult;
-        }
-        if (rawItem.expected !== undefined && typeof rawItem.expected !== 'string') {
-          return {
-            success: false,
-            error: `Invalid test_data[${index}].expected: expected string`,
-          } as GatewayToolResult;
-        }
-        normalizedItems.push({
-          input: rawItem.input,
-          ...(typeof rawItem.expected === 'string' ? { expected: rawItem.expected } : {}),
-        });
-      }
-      items = normalizedItems;
-    } else if (this.rawStore) {
-      const agentConfig = this.delegationManagerRef.getAgentConfig(agentId);
-      const connectors: string[] = (agentConfig?.connectors as string[]) ?? [];
-      if (connectors.length === 0) {
-        return {
-          success: false,
-          error: 'connector_unavailable: no connectors configured',
-        } as GatewayToolResult;
-      }
-      const allItems: Array<{ input: string }> = [];
-      const missingConnectors: string[] = [];
-      for (const conn of connectors) {
-        if (!this.rawStore.hasConnector(conn)) {
-          missingConnectors.push(conn);
-          continue;
-        }
-        const recent = this.rawStore.getRecent(conn, sampleCount);
-        for (const item of recent) {
-          allItems.push({ input: `[${item.type}] ${item.content}` });
-        }
-        if (allItems.length >= sampleCount) {
-          break;
-        }
-      }
-      if (allItems.length === 0) {
-        const detail =
-          missingConnectors.length > 0
-            ? `connector(s) not found: ${missingConnectors.join(', ')}`
-            : 'no recent data';
-        return {
-          success: false,
-          error: `connector_unavailable: ${detail}`,
-        } as GatewayToolResult;
-      }
-      items = allItems.slice(0, sampleCount);
-    } else {
-      return {
-        success: false,
-        error: 'connector_unavailable: rawStore not available',
-      } as GatewayToolResult;
-    }
-
-    // 2. Start validation session for agent_test
-    const testVer = this.sessionsDb ? getLatestVersion(this.sessionsDb, agentId) : null;
-    const testAgentVersion = testVer?.version ?? 0;
-    let testValSession: ValidationSessionRow | null = null;
-    try {
-      testValSession =
-        this.validationService?.startSession(agentId, testAgentVersion, 'agent_test', {
-          goal: `Test with ${items.length} items`,
-          customBeforeSnapshot: JSON.stringify({
-            schema_version: 1,
-            test_input_summary: items.map((i) => i.input.slice(0, 80)).join('; '),
-            sample_count: items.length,
-          }),
-        }) ?? null;
-    } catch (telemetryErr) {
-      securityLogger.warn('[Agent test telemetry] Failed to start validation session', {
-        agentId,
-        testAgentVersion,
-        error: telemetryErr instanceof Error ? telemetryErr.message : String(telemetryErr),
-      });
-    }
-
-    // 3. Log test_run start
-    let testRunId: number | null = null;
-    if (this.sessionsDb) {
-      try {
-        const row = logActivity(this.sessionsDb, {
-          agent_id: agentId,
-          agent_version: testAgentVersion,
-          type: 'test_run',
-          input_summary: `Testing with ${items.length} items`,
-          run_id: testValSession?.id,
-          execution_status: 'started',
-          trigger_reason: 'agent_test',
-        });
-        testRunId = row.id;
-        if (testValSession && this.validationService) {
-          this.validationService.recordRun(testValSession.id, { activityId: row.id });
-        }
-      } catch (telemetryErr) {
-        securityLogger.warn('[Agent test telemetry] Failed to persist startup activity', {
-          agentId,
-          testValSessionId: testValSession?.id ?? null,
-          error: telemetryErr instanceof Error ? telemetryErr.message : String(telemetryErr),
-        });
-        testValSession = this.cleanupValidationSessionOnTelemetryFailure(
-          testValSession,
-          telemetryErr,
-          'agent_test startup telemetry failed'
-        );
-        testRunId = null;
-      }
-    }
-
-    // 4. Delegate with a small concurrency limit to keep tests responsive
-    const results: Array<{ input: string; output?: string; error?: string }> = [];
-    const workerCount = Math.min(3, items.length);
-    let nextIndex = 0;
-    const workers = Array.from({ length: workerCount }, async () => {
-      for (;;) {
-        const currentIndex = nextIndex;
-        nextIndex++;
-        if (currentIndex >= items.length) {
-          return;
-        }
-        const item = items[currentIndex];
-        try {
-          const r = await this.executeDelegate({
-            agentId,
-            task: `Process this data:\n${item.input}`,
-          });
-          const rAny = r as Record<string, unknown>;
-          const output = r.success
-            ? String((rAny.data as Record<string, unknown>)?.response ?? '')
-            : undefined;
-          results[currentIndex] = {
-            input: item.input,
-            output,
-            error: r.success ? undefined : String(rAny.error ?? 'unknown'),
-          };
-        } catch (err) {
-          results[currentIndex] = { input: item.input, error: String(err) };
-        }
-      }
-    });
-    await Promise.all(workers);
-
-    // 5. Auto-score: pass/fail ratio
-    const passed = results.filter((r, index) => {
-      if (r.error) {
-        return false;
-      }
-      const expected = items[index]?.expected;
-      if (expected === undefined) {
-        return true;
-      }
-      return (r.output ?? '').trim() === expected.trim();
-    }).length;
-    const failed = results.length - passed;
-    const autoScore = results.length > 0 ? Math.round((passed / results.length) * 100) : 0;
-
-    if (this.sessionsDb && testRunId) {
-      try {
-        updateActivityScore(
-          this.sessionsDb,
-          testRunId,
-          autoScore,
-          {
-            total: results.length,
-            passed,
-            failed,
-            items: results.map((r, index) => ({
-              input: r.input.slice(0, 100),
-              result:
-                r.error ||
-                (items[index]?.expected !== undefined &&
-                  (r.output ?? '').trim() !== items[index]!.expected!.trim())
-                  ? 'fail'
-                  : 'pass',
-            })),
-          },
-          'completed'
-        );
-      } catch (telemetryErr) {
-        securityLogger.warn('[Agent test telemetry] Failed to persist test score', {
-          agentId,
-          testRunId,
-          autoScore,
-          totalResults: results.length,
-          error: telemetryErr instanceof Error ? telemetryErr.message : String(telemetryErr),
-        });
-      }
-    }
-
-    // 6. Finalize validation session with test metrics
-    const testDurationMs = Date.now() - startTime;
-    if (testValSession && this.validationService) {
-      try {
-        this.validationService.finalizeSession(testValSession.id, {
-          execution_status: 'completed',
-          metrics: {
-            duration_ms: testDurationMs,
-            completion_rate: results.length > 0 ? passed / results.length : 0,
-            auto_score: autoScore,
-          },
-          test_input_summary: items.map((i) => i.input.slice(0, 80)).join('; '),
-        });
-      } catch (telemetryErr) {
-        securityLogger.warn('[Agent test telemetry] Failed to finalize validation session', {
-          agentId,
-          testValSessionId: testValSession.id,
-          autoScore,
-          totalResults: results.length,
-          error: telemetryErr instanceof Error ? telemetryErr.message : String(telemetryErr),
-        });
-      }
-    }
-
-    return {
-      success: true,
-      data: {
-        test_run_id: testRunId,
-        agent_id: agentId,
-        results,
-        auto_score: autoScore,
-        duration_ms: testDurationMs,
-        validation_session_id: testValSession?.id ?? null,
-        ...(testRunId === null ? { warning: 'score_not_persisted' } : {}),
-      },
-    } as GatewayToolResult;
-  }
-
-  // ── Delegation ────────────────────────────────────────────────────────────
-
-  private async executeDelegate(input: {
-    agentId: string;
-    task: string;
-    background?: boolean;
-    skill?: string;
-  }): Promise<GatewayToolResult> {
-    const { agentId, task, background } = input;
-
-    // Resolve skill path safely — reject path traversal attempts
-    const resolveSkillPath = (skillName: string): string | null => {
-      if (!/^[a-zA-Z0-9_-]+$/.test(skillName)) {
-        return null;
-      }
-      const skillsDir = join(homedir(), '.mama', 'skills');
-      const resolved = resolve(skillsDir, `${skillName}.md`);
-      if (!resolved.startsWith(skillsDir)) {
-        return null;
-      }
-      return resolved;
-    };
-
-    if (!this.agentProcessManager || !this.delegationManagerRef) {
-      return { success: false, error: 'Multi-agent not configured' } as GatewayToolResult;
-    }
-
-    // Permission check using existing DelegationManager
-    // Default to 'conductor' when no agent context is set (e.g., MessageRouter path, audit cron)
-    // Conductor is the default agent and the only tier-1 agent that should delegate
-    const {
-      agentId: activeAgentId,
-      source: activeSource,
-      channelId: activeChannelId,
-    } = this.getActiveRouting();
-    const sourceAgentId = activeAgentId || 'conductor';
-    const check = this.delegationManagerRef.isDelegationAllowed(sourceAgentId, agentId);
-    if (!check.allowed) {
-      return {
-        success: false,
-        error: `Delegation denied: ${check.reason}`,
-      } as GatewayToolResult;
-    }
-
-    // Background delegation: fire-and-forget with async validation finalize
-    if (background) {
-      const source = activeSource || 'viewer';
-      const channelId = activeChannelId || 'default';
-
-      // Start validation session for background delegation
-      let bgAgentVersion = 0;
-      let bgValSession: ValidationSessionRow | null = null;
-      try {
-        const bgVer = this.sessionsDb ? getLatestVersion(this.sessionsDb, agentId) : null;
-        bgAgentVersion = bgVer?.version ?? 0;
-        bgValSession =
-          this.validationService?.startSession(agentId, bgAgentVersion, 'delegate_run') ?? null;
-
-        if (this.sessionsDb) {
-          const row = logActivity(this.sessionsDb, {
-            agent_id: agentId,
-            agent_version: bgAgentVersion,
-            type: 'task_start',
-            input_summary: task?.slice(0, 200),
-            run_id: bgValSession?.id,
-            execution_status: 'started',
-            trigger_reason: 'delegate_run',
-          });
-          if (bgValSession && this.validationService) {
-            this.validationService.recordRun(bgValSession.id, { activityId: row.id });
-          }
-        }
-      } catch (telemetryErr) {
-        securityLogger.warn('[Delegation telemetry] Background bootstrap failed', telemetryErr);
-        bgAgentVersion = 0;
-        bgValSession = this.cleanupValidationSessionOnTelemetryFailure(
-          bgValSession,
-          telemetryErr,
-          'background delegate bootstrap failed'
-        );
-      }
-
-      // Fire-and-forget: finalize validation when complete
-      const bgStartTime = Date.now();
-      void (async () => {
-        try {
-          const process = await this.agentProcessManager!.getProcess(source, channelId, agentId);
-          let delegationPrompt = this.delegationManagerRef!.buildDelegationPrompt(
-            sourceAgentId,
-            task
-          );
-          if (input.skill) {
-            const skillPath = resolveSkillPath(input.skill);
-            if (skillPath && existsSync(skillPath)) {
-              const skillContent = readFileSync(skillPath, 'utf-8');
-              delegationPrompt = skillContent + '\n\n---\n\n' + delegationPrompt;
-            }
-          }
-          const result = await process.sendMessage(delegationPrompt);
-          const durationMs = Date.now() - bgStartTime;
-
-          try {
-            if (this.sessionsDb) {
-              const row = logActivity(this.sessionsDb, {
-                agent_id: agentId,
-                agent_version: bgAgentVersion,
-                type: 'task_complete',
-                input_summary: task?.slice(0, 200),
-                output_summary: summarizeActivityOutput(result?.response),
-                duration_ms: durationMs,
-                run_id: bgValSession?.id,
-                execution_status: 'completed',
-                trigger_reason: 'delegate_run',
-              });
-              if (bgValSession && this.validationService) {
-                this.validationService.recordRun(bgValSession.id, {
-                  activityId: row.id,
-                  duration_ms: durationMs,
-                });
-              }
-            }
-          } catch (telemetryErr) {
-            securityLogger.warn(
-              '[Delegation telemetry] Background completion activity failed',
-              telemetryErr
-            );
-          }
-
-          try {
-            if (bgValSession && this.validationService) {
-              this.validationService.finalizeSession(bgValSession.id, {
-                execution_status: 'completed',
-                metrics: { duration_ms: durationMs },
-              });
-            }
-          } catch (telemetryErr) {
-            securityLogger.warn(
-              '[Delegation telemetry] Background completion finalize failed',
-              telemetryErr
-            );
-          }
-        } catch (err) {
-          const durationMs = Date.now() - bgStartTime;
-          try {
-            if (this.sessionsDb) {
-              const row = logActivity(this.sessionsDb, {
-                agent_id: agentId,
-                agent_version: bgAgentVersion,
-                type: 'task_error',
-                input_summary: task?.slice(0, 200),
-                error_message: String(err),
-                duration_ms: durationMs,
-                run_id: bgValSession?.id,
-                execution_status: 'failed',
-                trigger_reason: 'delegate_run',
-              });
-              if (bgValSession && this.validationService) {
-                this.validationService.recordRun(bgValSession.id, {
-                  activityId: row.id,
-                  duration_ms: durationMs,
-                });
-              }
-            }
-          } catch (telemetryErr) {
-            securityLogger.warn(
-              '[Delegation telemetry] Background failure activity failed',
-              telemetryErr
-            );
-          }
-          try {
-            if (bgValSession && this.validationService) {
-              this.validationService.finalizeSession(bgValSession.id, {
-                execution_status: 'failed',
-                error_message: String(err),
-                metrics: { duration_ms: durationMs },
-              });
-            }
-          } catch (telemetryErr) {
-            securityLogger.warn(
-              '[Delegation telemetry] Background failure finalize failed',
-              telemetryErr
-            );
-          }
-        }
-      })();
-
-      return {
-        success: true,
-        data: { agentId, background: true, message: 'Background task submitted' },
-      } as GatewayToolResult;
-    }
-
-    // Synchronous delegation with retry + backoff for resilience
-    const source = activeSource || 'viewer';
-    const channelId = activeChannelId || 'default';
-    const startTime = Date.now();
-
-    // Start validation session for this delegation
-    let agentVersion = 0;
-    let valSession: ValidationSessionRow | null = null;
-    try {
-      const ver = this.sessionsDb ? getLatestVersion(this.sessionsDb, agentId) : null;
-      agentVersion = ver?.version ?? 0;
-      valSession =
-        this.validationService?.startSession(agentId, agentVersion, 'delegate_run') ?? null;
-
-      if (this.sessionsDb) {
-        const row = logActivity(this.sessionsDb, {
-          agent_id: agentId,
-          agent_version: agentVersion,
-          type: 'task_start',
-          input_summary: task?.slice(0, 200),
-          run_id: valSession?.id,
-          execution_status: 'started',
-          trigger_reason: 'delegate_run',
-        });
-        if (valSession && this.validationService) {
-          this.validationService.recordRun(valSession.id, { activityId: row.id });
-        }
-      }
-    } catch (telemetryErr) {
-      securityLogger.warn('[Delegation telemetry] Validation bootstrap failed', telemetryErr);
-      agentVersion = 0;
-      valSession = this.cleanupValidationSessionOnTelemetryFailure(
-        valSession,
-        telemetryErr,
-        'delegate bootstrap failed'
-      );
-    }
-
-    const MAX_RETRIES = 3;
-    let lastError: Error | null = null;
-
-    for (let attempt = 0; attempt < MAX_RETRIES; attempt++) {
-      try {
-        const process = await this.agentProcessManager.getProcess(source, channelId, agentId);
-
-        let delegationPrompt = this.delegationManagerRef.buildDelegationPrompt(sourceAgentId, task);
-
-        // Inject skill content if specified
-        if (input.skill) {
-          const skillPath = resolveSkillPath(input.skill);
-          if (skillPath && existsSync(skillPath)) {
-            const skillContent = readFileSync(skillPath, 'utf-8');
-            delegationPrompt = skillContent + '\n\n---\n\n' + delegationPrompt;
-          }
-        }
-
-        // Inject channel history for fresh processes (no prior context)
-        const sessionId = process.getSessionId?.();
-        if (!sessionId || attempt > 0) {
-          try {
-            const { getChannelHistory } = await import('../gateways/channel-history.js');
-            const channelHistory = getChannelHistory();
-            if (channelHistory) {
-              const historyContext = channelHistory.formatForContext(channelId, '', agentId);
-              if (historyContext) {
-                delegationPrompt = `${historyContext}\n\n${delegationPrompt}`;
-              }
-            }
-          } catch {
-            // Channel history injection is best-effort
-          }
-        }
-
-        const result = await process.sendMessage(delegationPrompt);
-        const durationMs = Date.now() - startTime;
-
-        try {
-          if (this.sessionsDb) {
-            const row = logActivity(this.sessionsDb, {
-              agent_id: agentId,
-              agent_version: agentVersion,
-              type: 'task_complete',
-              input_summary: task?.slice(0, 200),
-              output_summary: summarizeActivityOutput(result.response),
-              duration_ms: durationMs,
-              run_id: valSession?.id,
-              execution_status: 'completed',
-              trigger_reason: 'delegate_run',
-            });
-            if (valSession && this.validationService) {
-              this.validationService.recordRun(valSession.id, {
-                activityId: row.id,
-                duration_ms: durationMs,
-              });
-            }
-          }
-        } catch (telemetryErr) {
-          securityLogger.warn('[Delegation telemetry] Completion activity failed', telemetryErr);
-        }
-
-        try {
-          if (valSession && this.validationService) {
-            this.validationService.finalizeSession(valSession.id, {
-              execution_status: 'completed',
-              metrics: { duration_ms: durationMs },
-            });
-          }
-        } catch (telemetryErr) {
-          securityLogger.warn('[Delegation telemetry] Completion finalize failed', telemetryErr);
-        }
-
-        return {
-          success: true,
-          data: { agentId, response: result.response, duration_ms: durationMs },
-        } as GatewayToolResult;
-      } catch (err) {
-        lastError = err instanceof Error ? err : new Error(String(err));
-        const isBusy = lastError.message.includes('busy');
-        const isCrash = lastError.message.includes('exited with code');
-
-        if (isCrash) {
-          this.agentProcessManager.stopProcess(source, channelId, agentId);
-        }
-
-        if (attempt < MAX_RETRIES - 1 && (isBusy || isCrash)) {
-          await new Promise((r) => setTimeout(r, this._retryDelayMs * (attempt + 1)));
-          continue;
-        }
-        break;
-      }
-    }
-
-    const failedDurationMs = Date.now() - startTime;
-    try {
-      if (this.sessionsDb) {
-        const row = logActivity(this.sessionsDb, {
-          agent_id: agentId,
-          agent_version: agentVersion,
-          type: 'task_error',
-          input_summary: task?.slice(0, 200),
-          error_message: lastError?.message,
-          duration_ms: failedDurationMs,
-          run_id: valSession?.id,
-          execution_status: 'failed',
-          trigger_reason: 'delegate_run',
-        });
-        if (valSession && this.validationService) {
-          this.validationService.recordRun(valSession.id, {
-            activityId: row.id,
-            duration_ms: failedDurationMs,
-          });
-        }
-      }
-    } catch (telemetryErr) {
-      securityLogger.warn('[Delegation telemetry] Failure activity failed', telemetryErr);
-    }
-
-    try {
-      if (valSession && this.validationService) {
-        this.validationService.finalizeSession(valSession.id, {
-          execution_status: 'failed',
-          error_message: lastError?.message,
-          metrics: { duration_ms: failedDurationMs },
-        });
-      }
-    } catch (telemetryErr) {
-      securityLogger.warn('[Delegation telemetry] Failure finalize failed', telemetryErr);
-    }
-
-    return {
-      success: false,
-      error: `Delegation to ${agentId} failed after ${MAX_RETRIES} attempts: ${lastError?.message}`,
-    } as GatewayToolResult;
   }
 
   /**
